move the metadata classes to a separate module
